requirements.txt
pymysql
flask
flask-babel
flask-sqlalchemy
flask-login
flask-admin